% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.above.count.R
\name{cols.above.count}
\alias{cols.above.count}
\title{Number of Columns with Value (at or) above (or below) Cutoff}
\usage{
cols.above.count(
  x,
  cutoff,
  or.tied = FALSE,
  na.rm = TRUE,
  below = FALSE,
  one.cut.per.col = FALSE
)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value.}

\item{cutoff}{The numeric threshold(s) or cutoff(s) to which numbers are compared.
Default is arithmetic mean of row (or mean of column, if one.cut.per.col = TRUE).
Usually one number.
Can be a vector of same length as number of rows (if one.cut.per.col=FALSE),
in which case each row can use a different cutoff.
Or, if one.cut.per.col = TRUE, then cutoff should be vector as long as the number of columns,
and each column is compared to its own cutoff.}

\item{or.tied}{Logical. Default is FALSE, which means we check if number in x is greater than the cutoff (>). If TRUE, check if greater than or equal (>=).}

\item{na.rm}{Logical value, optional, TRUE by default. Defines whether NA values should be removed before result is found. Otherwise result will be NA when a row has an NA value in any column.}

\item{below}{Logical. Default is FALSE. If TRUE, uses > or >= cutoff. If FALSE, uses < or <= cutoff.}

\item{one.cut.per.col}{Default is FALSE, which means there is just 1 cutoff same for all cases,
or cutoff is vector with one per row. If TRUE then cutoff is vector with 1 per column.}
}
\value{
Returns a vector the same size as the number of rows in x.
}
\description{
Find what number of columns have a value at or above some cutoff(s).
}
\details{
For a matrix with a few cols of related data, find what number of columns are at/above (or below) some cutoff(s).
Returns a vector of number indicating how many of the columns are at/above the cutoff(s).
Can be used in identifying places (rows) where some indicator(s) is/are at/above one or more cutpoints, threshold values.
}
\note{
Future work: these functions could have wts, na.rm, &
allow cutpoints or benchmarks as a vector (not just 1 number), & have benchnames.
}
\examples{
out <- cols.above.count(x<-data.frame(a=1:10, b=rep(7,10), c=7:16), cutoff=7)
out
out # default is or.tied=FALSE
out <- cols.above.count(data.frame(a=1:10, b=rep(7,10), c=7:16),
  cutoff=7, or.tied=TRUE, below=TRUE)
out
out <- cols.above.count(data.frame(a=1:10, b=rep(7,10), c=7:16) )
 # Compares each number in each row to the row's mean.
out
}
\seealso{
\code{\link{count.above}} \code{\link{pct.above}} \code{\link{pct.below}} to see, for each column, the count or percent of rows that have values above or below a cutoff.

\code{\link{cols.above.count}} \code{\link{cols.above.which}} \code{\link{cols.above.pct}} to see, for each row, the count or which or fraction of columns with numbers at/above/below cutoff.

colcounter_summary() colcounter_summary_cum() colcounter_summary_pct() colcounter_summary_cum_pct() tablefixed()

Other functions for above and below: 
\code{\link{cols.above.pct}()},
\code{\link{cols.above.which}()},
\code{\link{count.above}()},
\code{\link{count.below}()},
\code{\link{pct.above}()},
\code{\link{pct.below}()},
\code{\link{rows.above.count}()},
\code{\link{rows.above.pct}()},
\code{\link{rows.above.which}()},
\code{\link{rows.below.count}()},
\code{\link{rows.below.pct}()}
}
\concept{functions for above and below}
