% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.above.pct.R
\name{cols.above.pct}
\alias{cols.above.pct}
\title{Percent of Columns with Value at or above Cutoff}
\usage{
cols.above.pct(x, cutoff, or.tied = FALSE, na.rm = TRUE, below = FALSE)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value. Must have more than one row and one column?}

\item{cutoff}{The numeric threshold or cutoff to which numbers are compared. Default is arithmetic mean of row. Usually one number, but can be a vector of same length as number of rows, in which case each row can use a different cutoff.}

\item{or.tied}{Logical. Default is FALSE, which means we check if number in x is greater than the cutoff (>). If TRUE, check if greater than or equal (>=).}

\item{na.rm}{Logical, default TRUE. Should NA values be removed before analysis.}

\item{below}{Logical. Default is FALSE. If TRUE, uses > or >= cutoff. If FALSE, uses < or <= cutoff.}
}
\value{
Returns a vector the same size as the number of rows in x.
}
\description{
Find what percent of columns have a value at or above some cutoff.
}
\details{
For a matrix with a few cols of related data, find what percent of columns are at/above (or below) some cutoff.
Returns a vector of number indicating what percentage of the columns are at/above the cutoff.
Can be used in identifying places (rows) where some indicator(s) is/are at/above a cutoff, threshold value.
}
\examples{
out <- cols.above.pct(x<-data.frame(a=1:10, b=rep(7,10), c=7:16), cutoff=7)
out
out # default is or.tied=FALSE
out <- cols.above.pct(data.frame(a=1:10, b=rep(7,10), c=7:16),
  cutoff=7, or.tied=TRUE, below=TRUE)
out
out <- cols.above.pct(data.frame(a=1:10, b=rep(7,10), c=7:16) )
 # Compares each number in each row to the row's mean.
out

}
\seealso{
\code{\link[=count.above]{count.above()}}  \code{\link[=pct.above]{pct.above()}} \code{\link[=pct.below]{pct.below()}} to see, for each column, the count or percent of rows that have values above or below a cutoff.

\code{\link[=cols.above.count]{cols.above.count()}} \code{\link[=cols.above.which]{cols.above.which()}} \code{\link[=cols.above.pct]{cols.above.pct()}} to see, for each row, the count or which or fraction of columns with numbers at/above/below cutoff.

\code{\link[=colcounter_summary]{colcounter_summary()}} \code{\link[=colcounter_summary_cum]{colcounter_summary_cum()}} \code{\link[=colcounter_summary_pct]{colcounter_summary_pct()}} \code{\link[=colcounter_summary_cum_pct]{colcounter_summary_cum_pct()}}  \code{\link[=tablefixed]{tablefixed()}}
}
\author{
author
}
