% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.above.which.R
\name{cols.above.which}
\alias{cols.above.which}
\title{Does each Column have a Value at or above Cutoff(s)}
\usage{
cols.above.which(x, cutoff, or.tied = FALSE, below = FALSE)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value.}

\item{cutoff}{The numeric threshold or cutoff to which numbers are compared.
Default is arithmetic mean of row. Usually one number, but
can be a vector of same length as number of rows,
in which case each row can use a different cutoff.}

\item{or.tied}{Logical. Default is FALSE,
which means we check if number in x is greater than the cutoff (>).
If TRUE, check if greater than or equal (>=).}

\item{below}{Logical. Default is FALSE.
If TRUE, uses > or >= cutoff.
If FALSE, uses < or <= cutoff.}
}
\value{
Returns a logical matrix the same size as x. ** Note this is different than \code{\link{which}} --
  That function returns the positions of TRUE elements but this returns TRUE or FALSE for all elements.
}
\description{
Flag which cells are at or above some cutoff(s) or mean.
}
\details{
For a matrix with a few cols of related data, find which cells are at or above (or below) some cutoff.
  Returns a logical matrix, with TRUE for each cell that is at or above the cutoff.
  Can be used in identifying places (rows) where some indicator(s) is or are at or above a cutoff, threshold value.
}
\note{
Future work: these functions could have wts, na.rm, &
allow cutoffs or benchmarks as a vector (not just 1 number), & have benchnames.
}
\examples{
 out <- cols.above.which(x<-data.frame(a=1:10, b=rep(7,10), c=7:16), cutoff=7)
 out
 out # default is or.tied=FALSE
 out <- cols.above.which(data.frame(a=1:10, b=rep(7,10), c=7:16),
   cutoff=7, or.tied=TRUE, below=TRUE)
 out
 out <- cols.above.which(data.frame(a=1:10, b=rep(7,10), c=7:16) )
  # Compares each number in each row to the row's mean.
 out
}
\seealso{
\code{\link{count.above}} \code{\link{pct.above}} \code{\link{pct.below}} to see, for each column, the count or percent of rows that have values above or below a cutoff.

\code{\link{cols.above.count}} \code{\link{cols.above.which}} \code{\link{cols.above.pct}} to see, for each row, the count or which or fraction of columns with numbers at/above/below cutoff.

Other functions for above and below: \code{\link{cols.above.count}},
  \code{\link{cols.above.pct}}, \code{\link{count.above}},
  \code{\link{count.below}}, \code{\link{pct.above}},
  \code{\link{pct.below}}, \code{\link{rows.above.count}},
  \code{\link{rows.above.pct}},
  \code{\link{rows.above.which}},
  \code{\link{rows.below.count}},
  \code{\link{rows.below.pct}}
}
