% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2.R
\name{dir2}
\alias{dir2}
\title{Directory listing using wildcard search}
\usage{
dir2(x, ignore.case = TRUE, ...)
}
\arguments{
\item{x}{Query string that can use wildcards to search directory}

\item{ignore.case}{Logical, TRUE by default, optional. If FALSE, then this is case-sensitive.}

\item{...}{Optional other parameters passed to \code{\link[=dir]{dir()}}}
}
\value{
A directory listing.
}
\description{
Function to let you see directory listing using wildcard search syntax like '*.R'
}
\examples{
dir2('*.txt')
dir2('*.txt', path='~')
dir2()   # shows only files, not folders, if no x is specified.
dir2(path='~')  # shows only files, not folders, if no x is specified.
}
\seealso{
\code{\link[=dirdirs]{dirdirs()}} \code{\link[=dirr]{dirr()}}
}
