% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirdirs.R
\name{dirdirs}
\alias{dirdirs}
\title{Directory listing of R-related files/folders}
\usage{
dirdirs(path = ".", recursive = FALSE, ...)
}
\arguments{
\item{path}{Path as character string, optional. Default is current working directory.}

\item{recursive}{Logical value, optional, FALSE by default. Should subdirectories be shown.}

\item{...}{Optional other parameters passed to \code{\link{list.dirs}}}
}
\value{
A directory listing
}
\description{
Function to let you see directory listing of files/folders ending in r, R, or RData
}
\examples{
dirdirs()
}
\seealso{
\code{\link{dir2}} \code{\link{dirr}}
}

