% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirr.R
\name{dirr}
\alias{dirr}
\title{Directory listing of R-related files/folders}
\usage{
dirr(path = ".", ignore.case = TRUE, ...)
}
\arguments{
\item{path}{A file path string, optional, default is current working directory.}

\item{ignore.case}{Logical, TRUE by default, optional. If FALSE, then this is case-sensitive.}

\item{...}{Optional other parameters passed to \code{\link[=dir]{dir()}}}
}
\value{
A directory listing.
}
\description{
Function to let you see directory listing of files/folders ending in r, R, or RData
}
\seealso{
\code{\link[=dir2]{dir2()}} \code{\link[=dirdirs]{dirdirs()}}
}
