% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.files.R
\name{download.files}
\alias{download.files}
\title{Try to download one or more files}
\usage{
download.files(url, files, todir, silent = FALSE, overwrite = FALSE)
}
\arguments{
\item{url}{The url of folder with files to download, as character string}

\item{files}{A character vector of file names.}

\item{todir}{The folder where downloaded files will be placed, as a character string.}

\item{silent}{Logical, optional, FALSE by default. Prints a message using cat() if TRUE.}

\item{overwrite}{Optional, logical, FALSE by default. If FALSE, checks to see if file already exists in local folder and does not download if already exists.
But note that may cause problems if zero size file exists already due to earlier failed download.}
}
\value{
Returns vector of numbers, each being 1 or 0 or 2 to signify success or failure or no attempt because file already seems to exist locally.
}
\description{
Attempts to download files, given name(s) from specified url, saving them in specified folder.
Just a wrapper that Uses \code{\link{download.file}} which only downloads a single file.
}
\seealso{
\code{\link{download.file}}
}

