% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expand.gridMatrix.R
\name{expand.gridMatrix}
\alias{expand.gridMatrix}
\title{Similar to expand.grid, but returns a matrix not data.frame}
\usage{
expand.gridMatrix(x, y)
}
\arguments{
\item{x}{required vector}

\item{y}{required vector}
}
\value{
This function returns a matrix and tries to assign colnames based on the two input parameters. If they are variables, it uses those names as colnames. Otherwise it uses "x" and "y" as colnames.
}
\description{
This function is similar to \code{\link{expand.grid}}, in the sense that it returns a
  matrix that has 2 columns, one for each input, and one row per combination, cycling through the first field first.
  It differs from expand.grid in that this returns a matrix not data.frame, only accepts two parameters creating two columns, for now,
  and lacks the other parameters of expand.grid
}
\examples{
expand.gridMatrix(99:103, 1:2)
 zz <- 1:10; top <- 1:2
 expand.gridMatrix(zz, top)
}
\seealso{
\code{\link{expand.grid}}
}

