% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findArgs.R
\name{findArgs}
\alias{findArgs}
\title{Get the arguments of a function}
\usage{
findArgs(env, pattern)
}
\arguments{
\item{env}{e.g., 'package:analyze.stuff'}

\item{pattern}{search query as regular expression}
}
\value{
arguments
}
\description{
Get the arguments of a function as shown in help for \code{\link{grep}}
}
\examples{
findArgs("package:base", "warn")
}

