% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatcomma.R
\name{formatcomma}
\alias{formatcomma}
\title{Print numbers with commas and 0-2 decimal places}
\usage{
formatcomma(x, big.m = ",", nsmall = 2, drop0trailing = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{big.m}{Default is a comma at every three digits (1,000,000)}

\item{nsmall}{Default is 2 digits after the decimal (3.14)}

\item{drop0trailing}{Default is FALSE, which means zeroes after the last nonzero digit in the decimal portion are still shown (TRUE would mean 1.2 is shown, not 1.20)}

\item{...}{other parameters passed to \code{\link{format}}}
}
\value{
Same as \code{\link{format}} but with some preselected defaults
}
\description{
Wrapper for \code{\link{format}} making it easier to use with my typical settings
}
\examples{
cbind(mynum=formatcomma(c(1452345, 1.2, 4234.12345)))
 cbind(mynum=formatcomma(c(1452345, 1.2, 4234.12345), drop0trailing=FALSE))
}
\seealso{
\code{\link{format}}
}

