% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Geometric mean}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of numbers, required.}

\item{na.rm}{Logical value, optional, FALSE by default. If FALSE, result is NA if any of the values in x is NA. If TRUE, remove the NA values first.}
}
\value{
Returns a single number that is the geometric mean of the numbers in x.
}
\description{
Returns the geometric mean of a vector of numbers, which is the nth root of their product.
}
\details{
The geomean is one type of average, used in working with lognormal distributions, for example.
Is not as strongly influenced by extreme outliers as the arithmetic mean.
See \url{http://en.wikipedia.org/wiki/Geometric_mean} for many applications.
}
\examples{
geomean(c(4,9)) # is the square root of 4 * 9
}
\seealso{
\code{\link[=harmean]{harmean()}}  \code{\link[=mean]{mean()}}  \code{\link[=rms]{rms()}}
}
