% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installrequire.R
\name{installrequire}
\alias{installrequire}
\title{Require a list of packages, installing if necessary}
\usage{
installrequire(x, github)
}
\arguments{
\item{x}{vector of package names e.g., c("Hmisc", "data.table")}

\item{github}{optional vector of user slash package names e.g., "rstudio/shiny"}
}
\value{
none
}
\description{
Convenient way to specify packages to attach, and install any that are not already installed.
}
\details{
Uses \code{\link{require}} and if necessary uses \code{\link{install.packages}} or \code{\link{devtools::install_github}}
 If no parameters, prints an example. If github = 'ej' it installs several from github.
}
\examples{
installrequire()
}

