% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length2.R
\name{length2}
\alias{length2}
\title{Length of a list with or without NA values}
\usage{
length2(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector, required.}

\item{na.rm}{Logical value, optional, FALSE by default. Should NA values be left out of the count?}
}
\value{
Returns a single number.
}
\description{
Replacement for length(). Finds count of items like length(), but if set na.rm=TRUE then it doesn't count the items that are NA
}
\examples{
length2(c(1,2,3,NA))
length2(c(1,2,3,NA), na.rm=TRUE)
}

