% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefit.R
\name{linefit}
\alias{linefit}
\title{Add fit lines to a scatter plot}
\usage{
linefit(
  x,
  y,
  type = "b",
  cex = 4,
  show.lowess = TRUE,
  show.lm = TRUE,
  show.line = TRUE
)
}
\arguments{
\item{x}{x values, required}

\item{y}{y values, required}

\item{type}{passed through to \code{\link[graphics:lines]{graphics::lines()}} for the \code{\link[stats:lowess]{stats::lowess()}}}

\item{cex}{scaling for lowess}

\item{show.lowess}{Logical value, optional, TRUE by default. Defines if lowess is shown}

\item{show.lm}{Logical value, optional, TRUE by default. Defines if lm line is shown}

\item{show.line}{Logical value, optional, TRUE by default. Defines if should show abline(coef(line(x,y)))}
}
\value{
Provides a plot just as a side effect
}
\description{
Convenient wrapper for stats::lowess(), stats::lm(), and stats::coef(stats::line())
}
\details{
This function adds lines to a scatter plot, using
lines(\href{x,y}{stats::lowess()}),
abline(\href{y~x}{stats::lm()}), and
abline(\href{[stats::line()](x,y)}{stats::coef()})
DOESN'T SEEM TO WORK IF log = 'xy' was used in original plot()
NOTE: coef(line())  and lm()  give different results
}
\examples{
\dontrun{
    # see
#?lm  or  ?aov   or  ?glm
# ?line
require(graphics)
plot(cars)
(z <- line(cars))
abline(coef(z))
## Tukey-Anscombe Plot :
plot(residuals(z) ~ fitted(z), main = deparse(z$call))
# ?predict
# ?lowess
# ?scatterplot
#The scatterplot( ) function in the car package offers many enhanced features, including
#fit lines, marginal box plots, conditioning on a factor, and interactive point identification.
#Each of these features is optional.
# Enhanced Scatterplot of MPG vs. Weight
# by Number of Car Cylinders
 library(car)
 scatterplot(mpg ~ wt | cyl, data = mtcars,
             xlab = "Weight of Car", ylab = "Miles Per Gallon",
             main = "Enhanced Scatter Plot",
             labels = row.names(mtcars))
}
}
