% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linesofcode.R
\name{linesofcode}
\alias{linesofcode}
\title{Counts lines of source code in .R files of package source}
\usage{
linesofcode(folder = getwd(), packages, recursive = TRUE, sums = FALSE)
}
\arguments{
\item{folder}{Default is current working directory. **This is NOT the base path of the package itself!
It is the full path of the folder within which is a folder for each package of interest.
For example, folder= '~/Documents/R PACKAGES' works but folder= '~/Documents/R PACKAGES/mypkg' does not see the package called mypkg}

\item{packages}{Default is all found in folder. Can specify a subset of those by name as character vector.}

\item{recursive}{Default is TRUE, searches subfolders within specified folder.}

\item{sums}{Default is FALSE, but if TRUE it returns the count of .R files and lines of code for each package found.}
}
\value{
Returns a data.frame of results, with details depending on sums parameter.
 Also prints summary info if sums=FALSE, and returns detailed info.
}
\description{
This is just a way to summarize how many lines of code appear to be in the .R files in the folder that a package is built from.
}
\examples{
\dontrun{
 linesofcode(folder= '~/Documents/R PACKAGES', packages=c('analyze.stuff', 'proxistat') )
 x <- linesofcode(folder= '~/Documents/R PACKAGES')
 x[order(x$code), c('filename', 'package', 'code')]
 }
}

