% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logposneg.R
\name{logposneg}
\alias{logposneg}
\title{log10(x) if positive, 0 if 0, -log10(-x) if negative}
\usage{
logposneg(x)
}
\arguments{
\item{x}{numeric vector, required}
}
\value{
A numeric vector of same length as x
}
\description{
Function that transforms a vector of numbers x into log10(x) if positive, 0 if 0, -log10(-x) if negative,
useful for graphing something on a log scale when it has negative values.
This log scale expands outward from zero in both directions.
}

