% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/min.nonzero.R
\name{min.nonzero}
\alias{min.nonzero}
\title{Find minimum non-zero number(s)}
\usage{
\method{min}{nonzero}(mydf)
}
\arguments{
\item{mydf}{Required. Must be vector, matrix, or data.frame}
}
\value{
A number or vector of numbers
}
\description{
Returns minimum nonzero numbers in vector, matrix, or data.frame
}
\examples{
min.nonzero(-1:6)
min.nonzero(data.frame(a=0:10, b=1:11, c=c(0,1,NA)))
}

