% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.check.R
\name{na.check}
\alias{na.check}
\title{Basic info on each col of data.frame}
\usage{
na.check(mydat, min.text = FALSE)
}
\arguments{
\item{mydat}{Matrix or data.frame to examine. Cannot be a single vector currently.}

\item{min.text}{Logical, optional, defaults to FALSE. If TRUE, tries to find minimum of numbers stored as text? Slows it down.}
}
\value{
Returns a vector of results, one per col of df
}
\description{
Returns basic information on each field in a data.frame, like count of rows that are zero, negative,
NA, infinite, etc.
\cr\cr
Slow - work in progress
Leaves out logical, complex?, character, etc. cols
this version fails to handle fields that are factor class!
}
\examples{
na.check.m(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
}

