% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.check1.R
\name{na.check1}
\alias{na.check1}
\title{Basic info on each col of data.frame}
\usage{
na.check1(df, min.text = FALSE)
}
\arguments{
\item{df}{Matrix or data.frame to examine. Cannot be a single vector currently.}

\item{min.text}{Logical, optional, defaults to FALSE. If TRUE, tries to find minimum of numbers stored as text? Slows it down.}
}
\value{
Returns a vector of results, one per col of df
}
\description{
Returns basic information on each field in a data.frame, like count of rows that are zero, negative,
NA, infinite, etc.
\cr\cr
Slow - work in progress
Leaves out logical, complex?, character, etc. cols
}
\examples{
print(Sys.time()); x= na.check(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=FALSE)
print(Sys.time())
print(Sys.time()); x= na.check(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=TRUE)
print(Sys.time())
na.check(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
na.check1(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
na.check2(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
na.check3(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
}
\seealso{
\code{\link[matrixStats]{signTabulate}} \code{\link{minNonzero}} and experimental variations on na.check: \code{\link{na.check}} \code{\link{na.check1}} \code{\link{na.check2}} \code{\link{na.check3}}
}

