% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.check2.R
\name{na.check2}
\alias{na.check2}
\title{Basic info on each col of data.frame - testing faster way, but returns text}
\usage{
na.check2(df)
}
\arguments{
\item{df}{Matrix or data.frame to examine. Cannot be a single vector currently.}
}
\value{
Returns a vector of results, one per col of df
}
\description{
Returns basic information on each field in a data.frame, like count of rows that are zero, negative,
NA, infinite, etc.

Slow - work in progress
Leaves out logical, complex?, character, etc. cols
this version fails to handle fields that are factor class!?
}
\examples{
\dontrun{
system.time(x= na.check(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=FALSE) )
system.time(x= na.check2(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=TRUE) )
na.check(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
na.check2(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
}
}
\seealso{
\code{signTabulate} in \pkg{matrixStats} \code{\link{minNonzero}} and experimental variations on na.check: \code{\link{na.check}} \code{\link{na.check2}}
}

