% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/names2.R
\name{names2}
\alias{names2}
\title{Print names(data.frame) commented out for easy pasting into code}
\usage{
names2(x)
}
\arguments{
\item{x}{Data.frame, required}
}
\value{
Prints results
}
\description{
Uses cat() to print names of data.frame, but in a column with # before each.
Make it convenient to copy/paste into .R code as comments
}

