% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalized.R
\name{normalized}
\alias{normalized}
\title{Normalize raw scores as ratio of score to wtd mean}
\usage{
normalized(df, wts = NULL, na.rm = TRUE)
}
\arguments{
\item{df}{numeric Data.frame of one or more columns of values to be normalized, or matrix or vector to be coerced to data.frame}

\item{wts}{numeric Weights to use when computing weighted mean of given column, one weight per row in df (default=1) or per element of vector df. If omitted, default is unweighted mean.}

\item{na.rm}{logical Whether to exclude rows where weight or value or both = NA.}
}
\value{
matrix same size as df, but with all values in given column divided by weighted mean of that column
}
\description{
Provides a data.frame that takes the matrix or data.frame
and finds the weighted mean of each column and then divides each column of values by the column's weighted mean.
}
\details{
require(Hmisc) for \code{\link{scale}[Hmisc]}
}
\examples{
\dontrun{
 mydf_norm <- tbd
 ###
}
}
\seealso{
\code{\link{scale}[Hmisc]}
}

