% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/os.R
\name{os}
\alias{os}
\title{Windows, Mac, or other Unix?}
\usage{
os(x)
}
\arguments{
\item{x}{Optional query, must be among these:
'mac', 'apple', 'osx', 'darwin', 'win', 'windows', 'pc', 'microsoft', 'unix'}
}
\value{
If queried, returns TRUE or FALSE (or NA if query not recognized).
  If no query, returns 'win', 'mac', or 'unix'
}
\description{
Answers query about whether operating system is a certain type,
  or just reports type of operating system.
}
\examples{
os()
 if (os('mac')) {cat("Hi, I'm a Mac\\n")} else {cat('I am not a Mac\\n')}
}
\seealso{
\code{\link{get.os}} which is a bit more limited
}

