% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pct.below.R
\name{pct.below}
\alias{pct.below}
\title{Number or percent of rows (for each col) where value is below cutoff(s)}
\usage{
pct.below(df, benchmarks = "mean", benchnames = "cutoff", na.rm = FALSE,
  or.tied = FALSE, below = TRUE, wts = 1, of.what = "all")
}
\arguments{
\item{df}{Data.frame or matrix, required.}

\item{benchmarks}{Default is 'mean' but otherwise this must be a number or numeric vector of thresholds to compare values to.}

\item{benchnames}{Default is 'cutoff' and this string is used to create colnames for the results}

\item{na.rm}{Logical value, optional, TRUE by default. Defines whether NA values should be removed first. Otherwise result will be NA when any NA is in a col.}

\item{or.tied}{Logical, FALSE by default, reporting on those < cutoff. But, if or.tied=TRUE, this reports on those <= cutoff.}

\item{below}{Logical, TRUE by default, which counts how many are below cutoff (or tied if or.tied). If FALSE, counts how many are above (or tied with) cutoff.}

\item{wts}{Number or vector, default is 1. Length must be a factor of number of rows in df, so length(df[,1]) is an integer multiple of length(wts)  Applies weights to when counting how many.}
}
\value{
Returns a vector of numbers of length equal to number of columns in df.
}
\description{
Count the number or percent of rows (for each col of a data.frame) where the value is below some specified cutoff(s)
}
\details{
See \code{\link{pct.above}} for details, for which this is a wrapper.
}
\seealso{
\code{\link{rowMaxs}} \code{\link{rowMins}} \code{\link{colMaxs}} \code{\link{count.above}} \code{\link{cols.above.which}} \code{\link{cols.above.pct}}
}

