% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2.R
\name{pdf2}
\alias{pdf2}
\title{Overlay two simple histograms (pdf=probability density functions, not pdf filetype)}
\usage{
pdf2(x, binx, threshold, n = 100, colors = c("gray", "red"), ...)
}
\arguments{
\item{x}{Variable for histogram}

\item{binx}{Variable that defines subsets}

\item{threshold}{Variable that defines cutoff, so binx<threshold is plotted first, in one color, then binx>=threshold in other color is drawn over that}

\item{n}{Default is 100. n is just passed to the hist function.}

\item{colors}{Character vector length 2, with colors for first and second histogram}

\item{...}{Other named parameters sent to hist, such as main, xlab, ylab}
}
\value{
Just plots using hist.
}
\description{
Overlay two simple histograms, for data below vs above a cutoff
}
\examples{
  \dontrun{
  pdf2(
  log10(bg$proximity.npl),
  bg$bin.EJ.DISPARITY.proximity.npl.eo, 10,
  main ='npl for high top 10 pct vs low EJ', xlab='NPL score log10'
  )
  }
}
