% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmall.R
\name{rmall}
\alias{rmall}
\title{Help removing all objects from memory}
\usage{
rmall()
}
\value{
prints how to do that
}
\description{
A simple way to get a reminder of how to clear all objects from memory because I always forget how
}

