% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.R
\name{rms}
\alias{rms}
\title{Root Mean Square (RMS), or Quadratic Mean}
\usage{
rms(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of numbers, required.}

\item{na.rm}{Logical value, optional, FALSE by default. If FALSE, result is NA if any of the values in x is NA. If TRUE, remove the NA values first.}
}
\value{
Returns a single number
}
\description{
Returns the RMS, or quadratic mean of a vector of numbers.
}
\details{
The quadratic mean is one type of average. It is the square root of the arithmetic mean of the squares.
See \url{http://en.wikipedia.org/wiki/Root_mean_square} or \url{http://mathworld.wolfram.com/Root-Mean-Square.html} for many applications
}
\examples{
rms(c(1,2,4)) 
}
\seealso{
\code{\link[=geomean]{geomean()}}  \code{\link[=mean]{mean()}}  \code{\link[=harmean]{harmean()}}
}
