% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowMaxs.R
\name{rowMaxs}
\alias{rowMaxs}
\title{Returns the max value of each row of a data.frame or matrix}
\usage{
rowMaxs(df, na.rm = TRUE)
}
\arguments{
\item{df}{Data.frame or matrix, required.}

\item{na.rm}{Logical value, optional, TRUE by default. Defines whether NA values should be removed first. Otherwise result will be NA when any NA is in the given vector.}
}
\value{
Returns a vector of numbers of length equal to number of rows in df.
}
\description{
Returns maximum value of each row of a data.frame or matrix.
}
\details{
** Note than max() and min() default to na.rm=FALSE, but this function defaults to na.rm=TRUE because that just seems more frequently useful
\cr\cr
Also see the \pkg{matrixStats} package which uses some of the same names for some functions.
\cr\cr
Based on how min() and max() behave, return Inf or -Inf if no non-missing arguments to min or max respectively?
to call this and suppress that warning, use suppressWarnings( f(x) )
\cr\cr
NOTE: max() and min() & this function will handle character elements by coercing all others to character (see the help for Comparison \url{http://127.0.0.1:45798/help/library/base/help/Comparison})
which can be confusing -- e.g., note that min(c(8,10,'txt')) returns '10' not '8' and max returns 'txt'
\cr\cr
}
\examples{
blah <- rbind(NA, data.frame(a=c(0, 0:8), b=c(0.1+(0:9)), c=c(1:10), d=c(rep(NA, 10)), e=TRUE, f=factor('factor'), g='words', stringsAsFactors=FALSE) )
cbind(blah, min=rowMins(blah), max=rowMaxs(blah))
rbind(blah, min=colMins(blah), max=colMaxs(blah))
blah <- blah[ , sapply(blah, function(x) is.numeric(x) || is.logical(x)) ]
cbind(blah, min=rowMins(blah), max=rowMaxs(blah), mean=rowMeans(blah, na.rm=TRUE), sum=rowSums(blah, na.rm=TRUE))
rbind(blah, min=colMins(blah), max=colMaxs(blah), mean=colMeans(blah, na.rm=TRUE), sum=colSums(blah, na.rm=TRUE))
had been doing this:
mymin <- do.call(pmin, c(bg[,names.ej.pctile], na.rm=TRUE))
mymax <- do.call(pmax, c(bg[,names.ej.pctile], na.rm=TRUE))
now can do this:
mymin <- rowMins(bg[,names.ej.pctile])
mymax <- rowMaxs(bg[,names.ej.pctile])
}
\seealso{
\code{\link{rowMaxs}} \code{\link{rowMins}} \code{\link{colMaxs}} \code{\link{count.above}} \code{\link{cols.above.which}} \code{\link{cols.above.pct}}
}

