% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.below.count.R
\name{rows.below.count}
\alias{rows.below.count}
\title{Number or percent of rows (for each col) where value is below cutoff(s)}
\usage{
rows.below.count(
  df,
  benchmarks = "mean",
  benchnames = "cutoff",
  na.rm = TRUE,
  or.tied = FALSE,
  below = TRUE,
  wts = 1
)
}
\arguments{
\item{df}{Data.frame or matrix, required.}

\item{benchmarks}{Default is 'mean' but otherwise this must be a number or numeric vector of thresholds to compare values to.}

\item{benchnames}{Default is 'cutoff' and this string is used to create colnames for the results}

\item{na.rm}{Logical value, optional, TRUE by default. Defines whether NA values should be removed first. Otherwise result will be NA when any NA is in a col.}

\item{or.tied}{Logical, FALSE by default, reporting on those < cutoff. But, if or.tied=TRUE, this reports on those <= cutoff.}

\item{below}{Logical, TRUE by default, which counts how many are below cutoff (or tied if or.tied). If FALSE, counts how many are above (or tied with) cutoff.}

\item{wts}{Number or vector, default is 1. Length must be a factor of number of rows in df, so length(df[,1]) is an integer multiple of length(wts)  Applies weights to when counting how many.}
}
\value{
Returns a vector of numbers of length equal to number of columns in df.
}
\description{
Alias for \code{\link{count.below}}
}
\details{
See \code{\link{count.below}} for details, for which this is a wrapper.
}
\note{
Future work: these functions could have wts, na.rm, &
allow cutoffs or benchmarks as a vector (not just 1 number), & have benchnames.
}
\seealso{
\code{\link{count.above}} \code{\link{pct.above}} \code{\link{pct.below}} to see, for each column, the count or percent of rows that have values above or below a cutoff.

\code{\link{cols.above.count}} \code{\link{cols.above.which}} \code{\link{cols.above.pct}} to see, for each row, the count or which or fraction of columns with numbers at/above/below cutoff.

Other functions for above and below: 
\code{\link{cols.above.count}()},
\code{\link{cols.above.pct}()},
\code{\link{cols.above.which}()},
\code{\link{count.above}()},
\code{\link{count.below}()},
\code{\link{pct.above}()},
\code{\link{pct.below}()},
\code{\link{rows.above.count}()},
\code{\link{rows.above.pct}()},
\code{\link{rows.above.which}()},
\code{\link{rows.below.pct}()}
}
\concept{functions for above and below}
