% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff2.R
\name{setdiff2}
\alias{setdiff2}
\title{Differences between sets a and b}
\usage{
setdiff2(a, b)
}
\arguments{
\item{a}{Required vector}

\item{b}{Required vector}
}
\value{
Vector of elements
}
\description{
Returns the elements that in a or b but not in both (i.e., the differences between sets a and b)
}
\examples{
setdiff2(1:10, 3:12)
setdiff2(c('a','b','c'), c('b','c','d'))
}
\seealso{
\code{\link{setdiff}} which is a bit different
}
