% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifarray.R
\name{signifarray}
\alias{signifarray}
\title{Specify Significant Digits for Each Column}
\usage{
signifarray(dat, digits = 6)
}
\arguments{
\item{dat}{Required, matrix or numeric data.frame with the values to be rounded.}

\item{digits}{Optional, 6 by default. Can be a vector as long as the number of columns in dat, where each elements specifies the
number of significant digits to retain for numbers in the corresponding column of dat.}
}
\value{
Returns dat, but with numbers rounded based on digits parameter.
}
\description{
Given a matrix or numeric data.frame, round each column to a specified column-specific number of significant digits.
}
\examples{
 signifarray(matrix(rnorm(9*5), ncol=5), 1:5)
 signifarray(data.frame(a=rnorm(10), b=rnorm(10), c=rnorm(10)), 1:3)
}
\seealso{
\code{\link{signif}}
}

