% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{similar}
\alias{similar}
\title{See how closely numeric values match in 2 datasets}
\usage{
similar(a, b, tol = 99.99, na.rm = FALSE, shownames = TRUE)
}
\arguments{
\item{a}{Required first vector, data.frame, or matrix}

\item{b}{Required second vector, data.frame, or matrix}

\item{tol}{Number, 99.99 by default, specifying tolerance as a percentage 0-100, such that "similar" is defined as the two values being within 100-tol percent of each other.}

\item{na.rm}{Logical value, optional, FALSE by default. not implemented here yet. Should NA values be removed first, or compared and treated as NA matches NA.}

\item{shownames}{Logical value, optional, TRUE by default. Not used. Should names be shown in results?}
}
\value{
Data.frame showing what # of rows are "similar" in dataset a vs b, for each column.
}
\description{
Compare two vectors, matrices, or data.frames of numbers to see how often they are similar.
}
\details{
This function returns a matrix or vector showing how many rows in vector a are within 100-tol percent of the value in vector b.
May want to add a 3d case, where NA can match NA.
}
\examples{
 similar.p(1:10, (1:10) * 1.001 )
 similar.p(data.frame(x=1:10, y=101:110), data.frame(other=1.001*(1:10),
   other2=c(101:109, 110.01) ))
}
\seealso{
\code{\link[=similar.p]{similar.p()}}, \code{\link[=all.equal]{all.equal()}}, \code{\link[=identical]{identical()}}, \code{\link[=isTRUE]{isTRUE()}}, \code{\link[===]{==()}}, \code{\link[=all]{all()}}
}
