% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\title{Format a table in roxygen documentation of function in a package}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{data.frame required}

\item{...}{optional parameters passed through to lapply(df, format, ...)}
}
\value{
Returns text that can be pasted into documentation of a function or data in a package
}
\description{
modified version of func in help section on formatting in roxygen2 package
}
\examples{
  tabular(mtcars[1:5, 1:5])
  tabular(df = data.frame(a=7:16, b='stuff', c=999, d=c('blah','junk')) )
}
\seealso{
Help on formatting in roxygen2
}
