% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{tb}
\alias{tb}
\title{wrapper for table() that sorts by counts, decreasing}
\usage{
tb(x, useNA = "always", ...)
}
\arguments{
\item{x}{required, passed to table(x)}

\item{useNA}{default is 'always', passed to table()}

\item{...}{other parameters passed to table() -- cannot pass anything to cbind or sort like decreasing=FALSE}
}
\value{
like cbind
}
\description{
wrapper for table() that sorts by counts, decreasing
}
