% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.files.R
\name{unzip.files}
\alias{unzip.files}
\title{Unzip multiple zip files}
\usage{
unzip.files(
  zipfile,
  files = NULL,
  exdir = ".",
  unzip = "internal",
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{zipfile}{vector of names of files to unzip}

\item{files}{Optional, NULL by default which signifies all files in each zipfile will be extracted.
Otherwise, a list, with the nth element being a vector (length 1 or more) of character string names of files to extract from the nth zipfile.}

\item{exdir}{The directory to extract files to (the equivalent of unzip -d). It will be created if necessary.}

\item{unzip}{See help for unzip}

\item{overwrite}{Logical, optional, TRUE by default which means the local file is not overwritten if it already exists.}

\item{...}{Other arguments passed through to unzip}
}
\value{
Returns a list of the filepaths extracted to, from each zipfile. Names of list are the zip file names.
}
\description{
Wrapper for \code{\link[=unzip]{unzip()}} which unzips a single file.
}
