% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wtd.pctiles.fast.R
\name{wtd.pctiles.fast}
\alias{wtd.pctiles.fast}
\title{Show the values at 100 weighted percentiles}
\usage{
wtd.pctiles.fast(x, wts = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{Required numeric vector of values whose distribution you want to look at.}

\item{wts}{NULL by default, or vector of numbers to use as weights in Hmisc::wtd.quantile}

\item{na.rm}{Logical optional TRUE by default, in which case NA values are removed first.}
}
\value{
Returns a data.frame
}
\description{
Get a quick look at a weighted distribution by seeing the 100 values that are the weighted percentiles 1-100
}
\details{
Provides weighted percentiles without using \code{\link[Hmisc]{wtd.quantile}} \cr \cr
#  NOTE: THIS ONLY SHOWS PERCENTILES AND MEAN FOR THE VALID (NOT NA) VALUES !
Defining these types as type=1 and type="i/n" will create simple discontinuous quantiles, without interpolation where there are jumps in the values analyzed.
*** WARNING: Unless set type=1, the default type=7 in which case quantile() FUNCTION INTERPOLATES, WHICH ISN'T OBVIOUS IN EVERY DATASET!
use type=1 to avoid interpolation.
and pctiles() rounded results so interpolation would be even less apparent.\cr
The quantile function will NOT interpolate between values if type=1:\cr
quantile(1:12, probs=(1:10)/10, type=1)\cr
 10%  20%  30%  40%  50%  60%  70%  80%  90% 100% \cr
 2    3    4    5    6    8    9   10   11   12 \cr
###########################\cr
**** IMPORTANT ***\cr
###########################\cr
*** WARNING: The Hmisc::wtd.quantile function DOES interpolate between values, even if type='i/n'\cr
There does not seem to be a way to fix that for the Hmisc::wtd.quantile() function.
For example, \cr
Hmisc::wtd.quantile(1:12, probs=(1:10)/10, type='i/n', weights=rep(1,12))\cr
 10%  20%  30%  40%  50%  60%  70%  80%  90% 100% \cr
 1.2  2.4  3.6  4.8  6.0  7.2  8.4  9.6 10.8 12.0 \cr
}
\seealso{
\code{\link{pctiles}} \code{\link{pctiles.exact}} \code{\link{pctiles.a.over.b}} \code{\link{wtd.pctiles.exact}} \code{\link{wtd.pctiles}} \code{\link{wtd.pctiles.fast}}
}

