% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wtd.rowMeans.R
\name{wtd.rowMeans}
\alias{wtd.rowMeans}
\title{Weighted Mean of each Row - WORK IN PROGRESS}
\usage{
wtd.rowMeans(x, wts = 1, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{Data.frame or matrix, required.}

\item{wts}{Weights, optional, defaults to 1 which is unweighted, numeric vector of length equal to number of columns}

\item{na.rm}{Logical value, optional, TRUE by default. Defines whether NA values should be removed before result is found. Otherwise result will be NA when any NA is in a vector.}

\item{dims}{dims=1 is default. Not used. integer: Which dimensions are regarded as 'rows' or 'columns' to sum over. For row*, the sum or mean is over dimensions dims+1, ...; for col* it is over dimensions 1:dims.}
}
\value{
Returns a vector of numbers of length equal to number of rows in df.
}
\description{
Returns weighted mean of each row of a data.frame or matrix, based on specified weights, one weight per column.
}
\examples{
x=data.frame(a=c(NA, 2:10), b=rep(100,10), c=rep(3,10))
w=c(1.1, 2, NA)
cbind(x, wtd.rowMeans(x, w) )
}
\seealso{
\code{\link{wtd.colMeans}} \code{\link{rowMaxs}} \code{\link{rowMins}} \code{\link{colMins}}
}

