% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.os.R
\name{get.os}
\alias{get.os}
\title{Windows or Mac?}
\usage{
get.os()
}
\value{
Returns "win" or "mac" currently.
}
\description{
This function returns a character string "win" or "mac"
depending on which operating system is being used (that's all it does right now)
}

