% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.check.R
\name{na.check}
\alias{na.check}
\title{Basic info on each col of data.frame}
\usage{
na.check(df, min.text = FALSE)
}
\arguments{
\item{df}{Matrix or data.frame to examine. Cannot be a single vector currently.}

\item{min.text}{Logical, optional, defaults to FALSE. If TRUE, tries to find minimum of numbers stored as text? Slows it down.}
}
\value{
Returns a vector of results, one per col of df
}
\description{
Returns basic information on each field in a data.frame, like count of rows that are zero, negative,
NA, infinite, etc.
\cr\cr
Slow - work in progress
Leaves out logical, complex?, character, etc. cols
}
\examples{
\dontrun{
system.time(x= na.check(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=FALSE) )
system.time(x= na.check2(data.frame(a=-1:1e6, b='text', c=c(NA, 1, 2)), min.text=TRUE) )
na.check(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
na.check2(data.frame(a=-1:10, b='text', c=c(NA, 1, 2)))
}
}
\seealso{
\code{signTabulate} in \pkg{matrixStats} \code{\link{minNonzero}} and experimental variations on na.check: \code{\link{na.check}} \code{\link{na.check2}}
}

