% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlaps.R
\name{overlaps}
\alias{overlaps}
\title{Counts for Intersect, Union, etc. for Two Sets}
\usage{
overlaps(a, b)
}
\arguments{
\item{a}{Required vector, such as list of FIPS character codes.}

\item{b}{Required vector}
}
\value{
Returns a list of counts, formatted for viewing as a small table. See example.
}
\description{
This is just a convenient way to compare two sets (vectors) that overlap,
 to count how many are in each set, how many are in a not b, in b not a, in both, etc.
}
\examples{
overlaps( c('Selectric 251','Selectric 245'), c('Selectric 245','Selectric 255','Selectric 255'))
 \\donotrun{
          unique total
  in.a.only   1      1
  in.b.only   1      2
  overlap     1      2
  in.a        2      2
  in.b        2      3
  union       3      5
  in.one.only 2      3

  overlaps(ejanalysis::get.state.info()$ST, state.abb)
  data(fips.state, package='acs')
  overlaps(lead.zeroes(fips.state$STATE,2), ejanalysis::get.state.info()$FIPS.ST)
  data(fips.county, package='acs')
  overlaps(ejanalysis::get.county.info()$FIPS.COUNTY,
    paste(analyze.stuff::lead.zeroes(fips.county$State.ANSI,2),
    analyze.stuff::lead.zeroes(fips.county$County.ANSI,3), sep=''))
 }
}
\seealso{
\code{\link{setdiff2}}, \code{\link{sets}}, \code{\link[dplyr]{setops}}, \code{\link[plotrix]{intersectDiagram}}
}

