% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\title{Format a table in roxygen documentation of function in a package}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{data.frame required}

\item{...}{optional parameters passed through to lapply(df, format, ...)}
}
\value{
Returns text that can be pasted into documentation of a function or data in a package
}
\description{
taken from \code{\link[roxygen2]{formatting}} formatting \url{http://127.0.0.1:31032/help/library/roxygen2/doc/formatting.html}
}
\examples{
# cat(tabular(mtcars[1:5, 1:5]))
}
\seealso{
\code{\link[roxygen2]{formatting}}
}

